+++ app/src/main/java/com/travelbird/DatoRestaurante.kt
@@ -5,7 +5,7 @@ class DatoRestaurante {
     var nombre: String? = null
     var descripcion: String? = null
     var imagen: String? = null
-    var precioEstimado: Int? = null
+    var precioEstimado: Double? = null
     var ubicacion: String? = null
     var categorias: List<String>? = null
 
@@ -14,7 +14,7 @@ class DatoRestaurante {
         nombre: String?,
         descripcion: String?,
         imagen: String?,
-        precioEstimado: Int?,
+        precioEstimado: Double?,
         ubicacion: String?,
         categorias: List<String>?
     ) {
